separated components and added global.css
