+++ .storybook/main.ts
@@ -7,6 +7,11 @@ const config: StorybookConfig = {
     name: '@storybook/web-components-vite',
     options: {},
   },
+  previewHead: (head) => `
+  ${head}
+  <link rel="stylesheet" href="../src/styles/global.css" />
+  <link rel="stylesheet" href="../../../build/css/varsom.css" />
+`,
   docs: {
     autodocs: 'tag',
   },
