+++ index.html
@@ -9,10 +9,16 @@
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
     />
-    <link id="theme-link" rel="stylesheet" href="./src/varsom.css" />
-    <script type="module" src="/src/nve-button.ts"></script>
-    <script type="module" src="/src/nve-spinner.ts"></script>
-    <script type="module" src="/src/nve-icon.ts"></script>
+    <link rel="stylesheet" href="./build/css/varsom.css" />
+    <script
+      type="module"
+      src="/src/components/nve-button/nve-button.ts"
+    ></script>
+    <script
+      type="module"
+      src="/src/components/nve-spinner/nve-spinner.ts"
+    ></script>
+    <script type="module" src="/src/components/nve-icon/nve-icon.ts"></script>
   </head>
   <body>
     <div id="app"></div>
