+++ src/stories/NveButton.stories.ts
@@ -1,11 +1,18 @@
-import '../nve-button';
-import '../nve-spinner';
+import '../components/nve-button/nve-button';
+import '../components/nve-spinner/nve-spinner';
 import { NveButton } from './NveButton';
 
 export default {
   title: 'NveButton', // Title for the component in Storybook
   component: 'nve-button',
-  render: (args: { theme: string; variant: string; size: string; disabled: boolean; loading: boolean; outline: boolean; }) => NveButton(args),
+  render: (args: {
+    theme: string;
+    variant: string;
+    size: string;
+    disabled: boolean;
+    loading: boolean;
+    outline: boolean;
+  }) => NveButton(args),
   argTypes: {
     size: {
       control: { type: 'select' },
