+++ src/styles/global.css
@@ -1,0 +1,2 @@
+@import '@shoelace-style/shoelace/dist/themes/light.css';
+@import url('https://fonts.googleapis.com/css2?family=Source+Sans+Pro:wght@400;700&display=swap');
